--- Day 10: Adapter Array ---
I thought I could be clever with some graph theory maths because it's a DAG with a single topological sort, but I couldn't figure it out so I gave up and did the standard DP technique.

Also I realized that I had messed up the indexer range in day 9 but that I had gotten away with it because of my dataset.