--- Day 6: Tuning Trouble ---
So obviously there's a much more optimized way of handling this (by not repeating the same computations over and over and instead manually sliding the window over one by one), but there's something really satisfying about describing the problem in a lazy, stateless way. And hashing is pretty fast anyway, so it doesn't really matter given the small size of the input.