--- Day 15: Beacon Exclusion Zone ---
Thought about doing something cool like condensing connected cliques and then using the dual to find the "hole" in the graph, but then realized that that would only tell me the bounds of the potential areas, and I'd still have to do further math to figure out where the point might be with respect to the 3+ beacons that could surround it. So I used a SAT solver instead, because algebra is easier than geometry (probably).